move bilateral slice to contrib, test=develop
