+++ python/paddle/nn/functional/__init__.py
@@ -193,4 +193,3 @@ from .vision import shuffle_channel  #DEFINE_ALIAS
 from .vision import space_to_depth  #DEFINE_ALIAS
 from .vision import yolo_box  #DEFINE_ALIAS
 from .vision import yolov3_loss  #DEFINE_ALIAS
-from .vision import bilateral_slice  #DEFINE_ALIAS
