+++ python/paddle/nn/functional/vision.py
@@ -49,7 +49,6 @@ from ...fluid.layers import retinanet_detection_output  #DEFINE_ALIAS
 from ...fluid.layers import retinanet_target_assign  #DEFINE_ALIAS
 from ...fluid.layers import roi_perspective_transform  #DEFINE_ALIAS
 from ...fluid.layers import shuffle_channel  #DEFINE_ALIAS
-from ...fluid.layers import bilateral_slice  #DEFINE_ALIAS
 
 __all__ = [
     'affine_channel',
@@ -88,6 +87,5 @@ __all__ = [
     'shuffle_channel',
     'space_to_depth',
     'yolo_box',
-    'yolov3_loss',
-    'bilateral_slice'
+    'yolov3_loss'
 ]
